No need to use functor for non-exported parts
OptionMap is not exported, so there is no need to use a functor here.
Just use a simple module. What you are looking is in fact a default map
because this is the way you use it. A default map, is a map that returns
a default value if the element is not found.